backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
pandas==2.0.0
numpy==1.24.2
openpyxl==3.1.2
django-cors-headers==4.1.0